# allow embedding without renorm to be lowered as nvfuser index_select