The reason for the alternative lowering is that index_select can be handled by codegen, instead of lowering to aten.embedding.