@jjsjann123 2025-03-05
Running this with nvfuser embedding benchmark

On main
```
------------------------------------------------------------------------------------------------------------------- benchmark: 3 tests ------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                                     Min                 Max                Mean            StdDev              Median               IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_embedding_fwd_benchmark[executor='thunder'-variation='hf_qwen2']            100.9280 (1.0)      105.9200 (1.0)      102.4827 (1.0)      1.4943 (1.62)     102.1290 (1.0)      1.6310 (1.11)          2;1        9.7577 (1.0)          10           1
test_embedding_fwd_benchmark[executor='thunder'-variation='hf_mistral_nemo']     139.2950 (1.38)     143.9360 (1.36)     141.1164 (1.38)     1.4805 (1.61)     141.1355 (1.38)     2.1730 (1.48)          2;0        7.0863 (0.73)         10           1
test_embedding_fwd_benchmark[executor='thunder'-variation='hf_phi3']             169.3120 (1.68)     172.0650 (1.62)     170.6008 (1.66)     0.9212 (1.0)      170.4795 (1.67)     1.4710 (1.0)           4;0        5.8616 (0.60)         10           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

With changes in this PR

```
----------------------------------------------------------------------------------------------------------------- benchmark: 3 tests ----------------------------------------------------------------------------------------------------------------
Name (time in us)                                                                    Min                Max               Mean            StdDev             Median               IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_embedding_fwd_benchmark[executor='thunder'-variation='hf_qwen2']            41.1210 (1.0)      42.3050 (1.0)      41.5299 (1.0)      0.4080 (1.27)     41.3610 (1.0)      0.5420 (2.13)          3;0       24.0790 (1.0)          10           1
test_embedding_fwd_benchmark[executor='thunder'-variation='hf_mistral_nemo']     56.0330 (1.36)     57.0880 (1.35)     56.6116 (1.36)     0.3243 (1.01)     56.6240 (1.37)     0.3520 (1.38)          3;0       17.6642 (0.73)         10           1
test_embedding_fwd_benchmark[executor='thunder'-variation='hf_phi3']             67.7120 (1.65)     68.8630 (1.63)     68.3095 (1.64)     0.3207 (1.0)      68.3505 (1.65)     0.2550 (1.0)           3;2       14.6393 (0.61)         10           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```